refactor(maven): Rename ArgumentParser to CommandParser, move CommandParser to detectable/detectable/parser package.
