+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/clang/unit/CommandParserTest.java
@@ -33,17 +33,17 @@ import org.junit.jupiter.api.Test;
 
 import com.synopsys.integration.detectable.annotations.UnitTest;
 import com.synopsys.integration.detectable.detectables.clang.compilecommand.CompileCommand;
-import com.synopsys.integration.detectable.detectables.clang.compilecommand.ArgumentParser;
+import com.synopsys.integration.detectable.detectable.parser.CommandParser;
 
 @UnitTest
-public class ArgumentParserTest {
+public class CommandParserTest {
     @Test
     public void testGetCompilerCmd() {
         CompileCommand sampleCommand = new CompileCommand();
         sampleCommand.command = "g++ -DDOUBLEQUOTED=\"A value for the compiler\" -DSINGLEQUOTED='Another value for the compiler' file.c -o file.o";
 
-        ArgumentParser argumentParser = new ArgumentParser();
-        List<String> compilerCommand = argumentParser.parseCommand(sampleCommand, Collections.emptyMap());
+        CommandParser commandParser = new CommandParser();
+        List<String> compilerCommand = commandParser.parseCommand(sampleCommand, Collections.emptyMap());
 
         assertEquals("g++", compilerCommand.get(0));
     }
@@ -56,8 +56,8 @@ public class ArgumentParserTest {
         Map<String, String> optionOverrides = new HashMap<>();
         optionOverrides.put("-o", "/dev/null");
 
-        ArgumentParser argumentParser = new ArgumentParser();
-        List<String> result = argumentParser.parseCommand(sampleCommand, optionOverrides);
+        CommandParser commandParser = new CommandParser();
+        List<String> result = commandParser.parseCommand(sampleCommand, optionOverrides);
 
         for (String part : result) {
             System.out.printf("compiler arg: %s\n", part);
@@ -78,8 +78,8 @@ public class ArgumentParserTest {
         CompileCommand command = new CompileCommand();
         command.command = "/usr/bin/clang++-3.6 -DCMAKE_BUILD_TYPE=\\\"Debug\\\" -DCMAKE_CC_FLAGS=\"\\\" -ggdb -Wstrict-aliasing=2 -pedantic -fPIC --std=c11\\\"\"  -c ./pb.cc";
 
-        ArgumentParser argumentParser = new ArgumentParser();
-        List<String> result = argumentParser.parseCommand(command, Collections.emptyMap());
+        CommandParser commandParser = new CommandParser();
+        List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
 
         assertEquals(5, result.size());
         int i = 0;
@@ -95,8 +95,8 @@ public class ArgumentParserTest {
         CompileCommand command = new CompileCommand();
         command.command = "X=\"\\\" a  b\\\"\"";
 
-        ArgumentParser argumentParser = new ArgumentParser();
-        List<String> result = argumentParser.parseCommand(command, Collections.emptyMap());
+        CommandParser commandParser = new CommandParser();
+        List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
 
         assertEquals(1, result.size());
         int i = 0;
@@ -110,8 +110,8 @@ public class ArgumentParserTest {
         command.command = "X=\\\"'a' 'b'\\\"";
         command.file = "test.cc";
 
-        ArgumentParser argumentParser = new ArgumentParser();
-        List<String> result = argumentParser.parseCommand(command, Collections.emptyMap());
+        CommandParser commandParser = new CommandParser();
+        List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
         assertEquals(1, result.size());
         assertEquals("X=\\\"'a' 'b'\\\"", result.get(0));
     }
